Revert "[Feat] 장소 정보 불러오는 훅 작성 및 맵에서 로직 분리"
This reverts commit d3012cd0ca35ef10393c8b52391ef39fa19aae61.
