+++ src/pages/nearby-place/index.tsx
@@ -3,24 +3,10 @@ import Layout from '@/components/common/Layout';
 import Map from '@/components/nearby-place/Map';
 import Navigation from '@/components/common/Navigation';
 import BottomSheet from '@/components/nearby-place/BottomSheet';
-import useLocation from '@/hooks/useLocation';
-import { useEffect, useState } from 'react';
-
-declare global {
-  interface Window {
-    kakao: any;
-  }
-}
 
 function NearbyPlace() {
-  const [station, setStation] = useState('');
-  const { stationName } = useLocation();
-  useEffect(() => {
-    if (stationName) setStation(stationName);
-  }, [stationName]);
-
   return (
-    <Layout title={station} right={profile}>
+    <Layout title={'역삼역'} right={profile}>
       <Navigation />
       <div className="flex">
         <Map></Map>
